Add evolutionary algortihms

+++ super_tux_env.py
@@ -2,7 +2,6 @@ import gym
 import socket
 import struct
 import numpy
-import copy
 import subprocess
 import signal
 import select
@@ -14,54 +13,56 @@ class SuperTuxEnv(gym.Env):
 
     def __init__(self, proc_name, port, render):
         super(SuperTuxEnv).__init__()
-        self.inputs_count = 20
+        self.inputs_count = 10010
         self.actions_count = 6
-        self.outputs_count = 3
         self.port = port
         self.proc_name = proc_name
-        self.observation_space = spaces.Box(low=numpy.array([-325.0, -600.0,
-                                                             0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]),
-                                            high=numpy.array([325.0, 600.0,
-                                                              9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 9, 1]),
+
+        obs_array_low = [-325.0, -600.0, 0, 0, 0]
+        obs_array_low.extend([0] * 10000)
+        obs_array_high = [-325.0, -600.0, 1, 9, 9]
+        obs_array_high.extend([9] * 10000)
+
+        self.observation_space = spaces.Box(low=numpy.array(obs_array_low),
+                                            high=numpy.array(obs_array_high),
                                             dtype=numpy.float32)
         self.action_space = spaces.Discrete(self.actions_count)
         self.sock = None
         args = [self.proc_name, "--port", str(self.port)]
         if not render:
             args.extend(["--renderer", "null"])
-        self.subproc = subprocess.Popen(args=args)
+        self.subproc = subprocess.Popen(args=args, stdout=subprocess.PIPE)
 
-        # Observations
+        # Statistics
         self.play_time = .0
-        self.player_pos_x = .0
-        self.player_velocity_x = .0
-        self.player_velocity_y = .0
         self.player_run_record = .0
         self.player_env_record = .0
+        self.victories = 0
+        self.episodes = 0
+        self.visited_states = []
+
+        # Observations
+        self.player_pos_x = .0
+        self.player_pos_y = .0
+        self.player_velocity_x = .0
+        self.player_velocity_y = .0
         self.player_dead = .0
+        self.player_won = .0
         self.player_on_ground = .0
-        self.player_sensors = {
-            "x": .0,
-            "nx": .0,
-            "y": .0,
-            "ny": .0,
-            "xy": .0,
-            "xny": .0,
-            "lx": .0,
-            "lnx": .0,
-            "ly": .0,
-            "lny": .0,
-            "lxy": .0,
-            "lxny": .0,
+        self.player_box_sensors = []
+        self.player_ray_sensors = {
             "ground": .0,
-            "ground_front": .0
+            "ground_front": 0
         }
 
-    def __del__(self):
-        self.subproc.send_signal(signal.SIGINT)
+    def close(self):
+        if self.sock is not None:
+            self.sock.close()
+        if self.subproc is not None:
+            self.subproc.send_signal(signal.SIGINT)
 
     def step(self, action):
-        old_player_sensors = copy.deepcopy(self.player_sensors)
+        old_player_pos_x = self.player_pos_x
 
         # Send response corresponding to received action
         # Subtract 1 from action in order to allow model perform no action i.e. wait
@@ -77,43 +78,47 @@ class SuperTuxEnv(gym.Env):
             response[2] = 1
         self.sock.send(response)
         obs = self.get_observation()
+        # Save visited state
+        self.visited_states.append([int(self.play_time), int(self.player_pos_x), int(self.player_pos_y),
+                                    int(self.player_velocity_x), int(self.player_velocity_y)])
+        done = False
+        for sensor_name in self.player_ray_sensors:
+            sensor_read = self.player_ray_sensors[sensor_name]
+            self.player_ray_sensors[sensor_name] = 2 if sensor_read == 2 or sensor_read == 0 else 1
 
         # Calculate reward
-        enemy_nearby = False
         reward = .0
-        for sensor in self.player_sensors.items():
-            if sensor == 5:
-                enemy_nearby = True
-                reward -= 1
-        # Add acceleration info
-        reward += 0.2 if old_player_sensors["x"] == 1 and action == 2 else .0
-        reward += 1 if self.player_pos_x > self.player_run_record else .0
-        reward += 0.5 if self.player_velocity_x > 160 and self.player_on_ground == 1 and not enemy_nearby else .0
-        reward += 5 if self.player_sensors["ground_front"] == 2 and action == 3 else .0
-        reward += 5 if self.player_sensors["lxny"] == 5 and self.player_sensors["lx"] != 5 and action == 4 else .0
-
-        reward -= 1000 if self.player_dead >= 1.0 else .0
+        if self.player_ray_sensors["ground_front"] == 2:
+            reward += 5 if action == 3 else -5
+        reward += 1 if self.player_pos_x > old_player_pos_x else 0
+        reward += 2 if self.player_pos_x > self.player_run_record else -1
+        if self.play_time >= 120.0 or self.player_dead:
+            done = True
+            reward -= 1000
+        elif self.player_won:
+            done = True
+            reward += 1000
+            self.victories += 1
 
         # Set player record
         self.player_run_record = max(self.player_pos_x, self.player_run_record)
 
-        done = True if self.play_time >= 120.0 else False
         info = {}
 
-        if self.player_dead:
-            done = True
-
         if done:
+            self.episodes += 1
             self.player_env_record = max(self.player_env_record, self.player_run_record)
-            print(f"I'm done on {self.port}")
-            print(f"My record: {self.player_run_record}")
-            print(f"Environment record: {self.player_env_record}")
-            print(f"Dead: {self.player_dead}")
+            # print(f"I'm done on {self.port}")
+            # print(f"My record: {self.player_run_record}")
+            # print(f"Environment record: {self.player_env_record}")
+            # print(f"Dead: {self.player_dead}")
 
         return obs, reward, done, info
 
     def reset(self):
         self.player_run_record = 0
+        if self.player_won:
+            print(f"Player won! Total victories: {self.victories}")
         if self.sock is not None:
             self.sock.close()
         self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -124,52 +129,54 @@ class SuperTuxEnv(gym.Env):
     def render(self, mode='console'):
         pass
 
+    def receive_message(self):
+        chunks = []
+        bytes_received = 0
+        message_len = self.inputs_count * 4
+        while bytes_received < message_len:
+            is_ready = select.select([self.sock], [], [], 2)
+            if is_ready[0]:
+                chunk = self.sock.recv(message_len - bytes_received)
+            else:
+                self.sock.close()
+                self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+                self.sock.connect(("localhost", self.port))
+                return self.receive_message()
+            if chunk == b'':
+                raise RuntimeError("socket connection broken")
+            chunks.append(chunk)
+            bytes_received = bytes_received + len(chunk)
+        return b''.join(chunks)
+
     def get_observation(self):
         self.sock.setblocking(False)
-        is_ready = select.select([self.sock], [], [], 5)
-        if is_ready[0]:
-            read_buffer = self.sock.recv(self.inputs_count * 4)
-        else:
-            print(f"recv timeout on {self.port}")
-            print(f"Trying to reconnect on {self.port}...")
-            self.sock.close()
-            self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            self.sock.connect(("localhost", self.port))
-            return self.get_observation()
+
+        read_buffer = self.receive_message()
 
         if len(read_buffer) > 0:
             self.play_time = struct.unpack_from("f", read_buffer, 0)[0]
             self.player_pos_x = struct.unpack_from("f", read_buffer, 4)[0]
-            self.player_velocity_x = struct.unpack_from("f", read_buffer, 8)[0]
-            self.player_velocity_y = struct.unpack_from("f", read_buffer, 12)[0]
-            self.player_sensors["x"] = struct.unpack_from("f", read_buffer, 16)[0]
-            self.player_sensors["nx"] = struct.unpack_from("f", read_buffer, 20)[0]
-            self.player_sensors["y"] = struct.unpack_from("f", read_buffer, 24)[0]
-            self.player_sensors["ny"] = struct.unpack_from("f", read_buffer, 28)[0]
-            self.player_sensors["xy"] = struct.unpack_from("f", read_buffer, 32)[0]
-            self.player_sensors["xny"] = struct.unpack_from("f", read_buffer, 36)[0]
-            self.player_sensors["lx"] = struct.unpack_from("f", read_buffer, 40)[0]
-            self.player_sensors["lnx"] = struct.unpack_from("f", read_buffer, 44)[0]
-            self.player_sensors["ly"] = struct.unpack_from("f", read_buffer, 48)[0]
-            self.player_sensors["lny"] = struct.unpack_from("f", read_buffer, 52)[0]
-            self.player_sensors["lxy"] = struct.unpack_from("f", read_buffer, 56)[0]
-            self.player_sensors["lxny"] = struct.unpack_from("f", read_buffer, 60)[0]
-            self.player_sensors["ground"] = struct.unpack_from("f", read_buffer, 64)[0]
-            self.player_sensors["ground_front"] = struct.unpack_from("f", read_buffer, 68)[0]
-            self.player_dead = struct.unpack_from("f", read_buffer, 72)[0]
-            self.player_on_ground = struct.unpack_from("f", read_buffer, 76)[0]
+            self.player_pos_y = struct.unpack_from("f", read_buffer, 8)[0]
+            self.player_velocity_x = struct.unpack_from("f", read_buffer, 12)[0]
+            self.player_velocity_y = struct.unpack_from("f", read_buffer, 16)[0]
+            self.player_dead = struct.unpack_from("f", read_buffer, 20)[0]
+            self.player_won = struct.unpack_from("f", read_buffer, 24)[0]
+            self.player_on_ground = struct.unpack_from("f", read_buffer, 28)[0]
+            self.player_ray_sensors["ground"] = struct.unpack_from("f", read_buffer, 32)[0]
+            self.player_ray_sensors["ground_front"] = struct.unpack_from("f", read_buffer, 36)[0]
 
-            # Treat holes the same way as spikes
-            self.player_sensors["ground"] = 2 if self.player_sensors["ground"] == 0 else self.player_sensors["ground"]
-            self.player_sensors["ground_front"] = 2 if self.player_sensors["ground_front"] == 0 else self.player_sensors["ground"]
-            # Ignore all collectibles (e.g. Coins)
-            for sensor in self.player_sensors:
-                if self.player_sensors[sensor] == 8:
-                    self.player_sensors[sensor] = 0
+            self.player_box_sensors.clear()
+            for i in range(10000):
+                self.player_box_sensors.append(struct.unpack_from("f", read_buffer, 40 + (i * 4))[0])
 
-        return numpy.array([self.player_velocity_x, self.player_velocity_y, self.player_sensors["x"],
-                            self.player_sensors["nx"], self.player_sensors["y"], self.player_sensors["ny"],
-                            self.player_sensors["xy"], self.player_sensors["xny"], self.player_sensors["lx"],
-                            self.player_sensors["lnx"], self.player_sensors["ly"], self.player_sensors["lny"],
-                            self.player_sensors["lxy"], self.player_sensors["lxny"], self.player_sensors["ground"],
-                            self.player_sensors["ground_front"], self.player_on_ground]).astype(numpy.float32)
+        obs = [self.player_velocity_x, self.player_velocity_y, self.player_on_ground,
+               self.player_ray_sensors["ground"], self.player_ray_sensors["ground_front"]]
+        obs.extend(self.player_box_sensors)
+
+        return numpy.array(obs).astype(numpy.float32)
+
+    def get_visited_states(self):
+        state_tuples = []
+        for state in self.visited_states:
+            state_tuples.append(tuple(state))
+        return set(tuple(state_tuples))
